Add new factory function to configure ArgumentParser
This adds a new function to the wordsearch module to construct an
ArgumentParser (to be used to read command line arguments and provide
the parsed data to the program).

The function should return a configured ArgumentParser, but the first
step is to ensure that it does actually return a parser and not None.