Add method to look up valid moves
As we iterate through each position, it is helpful to be able to
determine if searching in a particular direction is possible or not. The
purpose of the get_valid_moves method is to look in each direction and
see if moving in that direction would move out of bounds or not. If it
stays in bounds, then the position is valid and included in the return
value. Otherwise, it is omitted.