Add find method to Puzzle class
The find method builds on the methods already available in the Puzzle
class by providing a means of searching for a given word in the puzzle.
It makes use of the all_positions generator to move through the board
one space at a time, the get_valid_moves to find all of the valid moves
at each position (with respect to the length of the word being
searched).

Then it uses get_characters to retrieve the characters in each range
from the current position to points given by get_valid_moves. Each
result is compared with the query word, and if they match, then the
positions of each character are returned to the caller.