Add find_all method to Puzzle class
To keep things simple for users of the Puzzle class, I added a find_all
method that wraps the find method but makes it work by returning the
results for a list of words instead of a single word. Since a typical
word search puzzle has more than one word to search for, this makes it
so that all words in the list can be found with a single method
invocation.