Add format_results function
The format_results function, aside from just printing whatever is in the
results dict, it actually has to reverse the x and y values for each
position because the library works with locations in the format (y,x),
or more clearly, (row, column), but the expected output is an (x, y)
pair. Since this only matters for the output, I decided to handle it as
the output string is being built.