Address invalid-name errors from pylint
After running pylint, it suggested that the use of single letter
variables x and y are not valid names. I agree with this in general, so
I improved the names of several instances, but I also wanted to keep
things concise, so where it was appropriate I added ignore statements to
allow it explicitly.